Allow return/set/update w/ a non-pljava SQLXML
Add to SQLXMLImpl the ability to accept a readable SQLXML object
that isn't an instance of SQLXMLImpl, and copy its content to
a new SQLXMLImpl for PostgreSQL's use.

Calls getSource(null) on the foreign object, giving it the choice
of what sort of Source to return; will accept a StreamSource,
SAXSource, StAXSource, or DOMSource. Does not yet do anything to
work around the possible case of a StreamSource that isn't in the
server encoding; encoding errors will be thrown.

In passing, arrange for StAXResultAdapter not to mess up element names.