Add API to adjust OWASP-suggested parser settings
XML parsers supplied in the Java runtime environment have many behavior
settings that may be appropriate when all input documents are of known
provenance, but are best changed to more restrictive settings if input
from unknown sources might be processed. The recommendations of the
Open Web Application Security Project naturally lean toward making the
settings restrictive by default, with a way for an application to relax
them when it has more trust in the input.

https://cheatsheetseries.owasp.org/cheatsheets/XML_External_Entity_Prevention_Cheat_Sheet.html#java

However, the recommended settings are so restrictive (such as rejecting
with an error any document with a DTD, even a purely internal one), that
it would be madness to impose them without a simple way for application
code to selectively relax them. And the Java APIs make tweaking those
settings insanely difficult. Most (but not all!) of the things that can
be set are identified by URIs, the exact spellings of which have to be
gleaned from multiple sources (or copied, cargo-cult-like, from the
OWASP cheat sheet), some have been spelled more than one way in
different supplied parsers or versions, and the details vary between the
major XML APIs (SAX, DOM, StAX, for starters) in the out-of-the-box Java
runtime.

Part of the appeal of JDBC's SQLXML interface is that it hides a lot of
the fiddly details needed to configure Java's different styles of XML
parser. You call getSource(SAXSource.class) and you get a SAXSource,
or getSource(StAXSource.class) and you get a StAXSource, and so for
DOMSource, all ready to use in each case. A side effect of that hiding
is that in two out of three of those cases, you don't have access to the
objects that let you adjust the settings. In the case of a SAXSource,
you can retrieve the object, but again, you change the settings by
throwing obscure mystery URIs at it until you find the ones it accepts.

So a simpler API is needed. In keeping with the extensible way SQLXML's
getSource() and getResult() are designed, you can call getSource and
pass, say, not SAXSource.class, but Adjusting.XML.SAXSource.class, and
you get back something with a little fluent API that lets you say
.allowDTD(true).externalGeneralEntities(true).xIncludeAware(true).get()
and the final get() returns an ordinary SAXSource with the desired
settings.

For now, this interface only allows settings on getSource(). Something
similar will be needed on setResult() also (if nothing else, to provide
a way to relax the settings of the verifier that eavesdrops on output
being written).