Ok, about that hex/base64 business
There's an earlier-overlooked subtlety in the spec. The algorithms
"Mapping SQL data types to XML Schema data types" and "Determination
of an XQuery formal type notation" have disjoint sets of callers, with
the latter being called to type parameters and context items passed in
to XMLQUERY and therefore XMLTABLE (and also, in the spec, the values
passed to XMLDOCUMENT and XMLPI, but not to other node constructors).
The former method is used in other places where the types need to be
mapped.

The callers of the determine-formal-type method correspond here to
callers of typeXS(), and the callers of the mapping-SQL-types method
to the callers of typeXT() (the T now seems mnemonic of "target" type).

Weirdly, the mapping-SQL-types method (and therefore now typeXT) is
specified to honor the xmlbinary setting (hex or base64) currently
in effect, but the determine-formal-type method is not: it's specified
instead to always declare xs:hexBinary as the Schema type for an SQL
binary string.

There could be an editorial mistake here needing to be clarified,
because in the rules for XMLQUERY, after declaring the formal type
statically with determine-formal-type (therefore as hex for a binary
string, irrespective of the xmlbinary setting), the value to be bound
comes from an application of XMLCAST (which uses mapping-SQL-types and
therefore honors the xmlbinary setting), and that could lead to
violating type safety by binding an xs:base64Binary where the declared
type is xs:hexBinary.

As a saving grace (which will have to do until some clarification can
be obtained on the spec), the mapJDBCofSQLvalueToXdmAtomicValue method
here was already ignoring its 'encoding' parameter and producing
whichever type (hex or base64) had been statically declared. For now,
it'll just keep right on doing that.