Now with Stream and SAX Results also
The adjustments offered on Result are, so far, the ones offered
on Source, and chiefly so there will be a way to adjust any
implicitly-created parser used to verify the content written to
the Result. (A couple additional methods provide a way to influence
whether a StreamResult produces an OutputStream or a Writer).

The SAX version is essentially decorative (for now, anyway, a
verifying parser isn't created for a SAX result), and StAX and
DOM versions are not even provided.