Expose the magic via Adjusting.XML.SourceResult
There is enough magic involved in carefully copying a 'foreign'
SQLXML instance to a native one that it should be exposed through
a new variety of Adjusting.XML.Result, so a caller can adjust any
parser restrictions if needed.

There might only be a 'parser' with restrictions to adjust if
the foreign SQLXML instance chooses to expose its content as
a StreamResult. For content presented in any already-parsed
form, the adjustment methods will be no-ops.

In passing, factor out the parsing adjustment methods to a new
interface Adjusting.XML.Parsing inherited by both Source and Result.