Testing ...
Already a couple unwelcome discoveries:

The SAX and DOM serializers will replace a character with a reference
if it is unrepresentable in the output encoding, but only in places
where a reference is acceptable. Elsewhere, they simply try to write
the character, through an encoder that will silently replace it with
a ?. Hardly acceptable for database infrastructure. It feels wasteful
to instantiate a CharsetEncoder explicitly and make them use that, but
it does succeed in getting exceptions rather than silent information
loss for the unrepresented characters.

The StAX serializer is less fastidious and will replace the character
with a reference even in a text node that is outside of any element.
Technically, that violates the letter of the XML spec (and in fact the
StAX parser will not accept the result!), but at least it is not
information-losing, and it is closer to the spirit of SQL/XML with its
XML(CONTENT) type.

Another headache: the use of a StAX parser in useWrappingElement
turns out to be treacherous; it will throw an exception if it
encounters a character reference before the first element. So much
for using it to decide whether a wrapping element is needed. Will
useWrappingElement need to be hand-rolled? Ugh.

This commit adds an explicit CharsetEncoder for SAXResult and
DOMResult to protect against silent data corruption. StAXResult
is left to do the more-XML(CONTENT)-friendly thing it does.