--release. It's not just a good idea.
When building with Java 9 or later, it is not only possible
but imperative to drop the older -source/-target options and
use --release instead, because of minor API tweaks such as
Java 9 ByteBuffer overriding Buffer methods that it didn't
in Java 8, narrowing their return types from Buffer to ByteBuffer.
If not prevented with --release, that will compile unchanged
Java 6-compatible source code into classes that give
NoSuchMethodError in a pre-Java-9 runtime.

So, if the build JDK is new enough to understand --release, then
the maven-compiler-plugin needs to be new enough to supply it,
and the right target release needs to be supplied.

The Maven pommery needed to cover the various cases may sap
irreversibly a small amount of light from the soul.