Work around Java 9+ XMLEventReader regression
An XMLEventWriter can be created over an XMLStreamWriter by
wrapping the latter in a StAXResult and using XMLOutputFactory's
createXMLEventWriter(Result) method. In Java 8 and earlier, that
will work for any underlying XMLStreamWriter object, but a
regression in Java 9 means you will only get a ClassCastException
for your troubles, if the XMLStreamWriter is anything but the
non-exported internal one from the Java runtime itself, which
application code has no way to extend.

Reported, but still in the notorious Java-bug-system limbo where
no number is assigned and no reply is heard for nobody knows
how long. "Internal review ID" 9062000, for all the good that
does anyone.