Let StAX copies preserve empty-element-ness
A pleasant side effect of having to reimplement XMLEventWriter
to XMLStreamWriter bridging is the chance to get the distinction
between empty elements and bare start/end element pairs right.
XMLStreamWriter has distinct methods for those, but there is no
obvious way to preserve the distinction at the XMLEventWriter
level.

There's a non-obvious way, though, if the StartElement and
EndElement events have parser-supplied Locations; if the
locations are the same (and not both all unknown), it's an
empty element.