Support column values that aren't atomic types
Away with the "fully general rules for returning XMLTABLE output
are not yet implemented" message.

Also, therefore, away with the stopgap casts sprinkled all through
the examples in the docs.

There remains a data() around the @id bare-attribute example, because
so help me, the spec seems to require err:XPTY0004 if an actual bare
attribute is returned.