Implement the missing casting of atomic outputs
Remove the "Casting AV to BV is not yet implemented" message
in XMLTABLE, and the accompanying need to write explicit casts
everywhere in the column expressions for atomic column types.

Still to do: XMLCAST General Rules 4(i)(vii) and (viii).
Also, the handling of column values that aren't atomic types.
For now, the "fully general rules for returning XMLTABLE output are
not yet implemented" message is still there.

The S9 code now uses Java 8 lambdas, which means that from now, when
built with the saxon-examples profile enabled, the examples
  -- all the examples, not just S9 :( --
will be built with Java 8 specified for source and target, and will
not run on earlier VMs.