Use Configuration.compileRegularExpression method
... rather than delving into Saxon's innards to construct a
specific class.

Beyond better API hygiene, this provides a little Easter egg
of sorts: Saxon's compileRegularExpression method allows a
non-standard ";j" suffix on the flags argument, and then builds
a Java regular expression instead of an XQuery one. Flagrantly
nonconformant, but could be handy at times, when some feature
that's in Java regular expressions and not in XQuery ones might
be helpful.

One good Easter egg deserves another, so detect when the Java-regex
egg has been used, and then allow TRANSLATE_REGEX to replace
zero-length matches rather than raising 2201U "attempt to replace
a zero-length string" as the SQL spec demands. XQuery regexes
don't have lookahead/lookbehind operators, but Java ones do, and so
can be constructed to match empty strings in perfectly sensible
and deliberate ways.

However, it's still tricky. The way TRANSLATE_REGEX is specified to
work, the RE is first applied to the whole input string, and then
for every substring that was matched as a replacement candidate, the
same RE is applied again, in the replacement step, to that substring
in isolation (a completely empty string, if that's what the pattern
matched). So if a Java pattern like (?<=c)(?=b) matched the empty
spot between c and b in 'fooabcbar', that pattern will not match
again in the replacement step, when the left and right context are
missing. A pattern that uses negative lookaheads/behinds to identify
its replacement spot will work fine, as will a pattern modified to
match EITHER the desired spot in context OR a completely empty (\A\z)
string.