Revise context, SPI, Invocation handling for SRFs
Some long-standing behavior in Type_invokeSRF needs a lot more
documentation and a bit of revision. Document that the C Invocation
structure is new for each call in the SFRM_ValuePerCall protocol,
but the Java Invocation instance is magically made to persist for
the protocol's duration.

Clarify that something like that is being done for the SPI connection
also. I am far from convinced that's even desirable, but it is thirteen
years old now, and changing it radically will demand more analysis.
However, it does need to handle more patterns of SPI use by the called
function without restoring a defunct memory context on popInvocation.

Simple fix is to do away with the "PL/Java row context", which seems to
serve no purpose: it is a context that gets reset for each row call, but
so is the one supplied as CurrentMemoryContext by the executor anyway,
and that one is still valid when our popInvocation happens.