Whitespace / tidying in SQLXMLImpl
Found one misindented line in saxCopy, and moved the whole method
nearer the only two places that use it, for ease of reading. This
actually does change what class it's a member of (now XMLCopier
rather than SQLXMLImpl), but that's downright appropriate, and it has
never appeared the other way in a release, so it's a trivial change
at this point. Also some off line-wrapping in a comment. Prune three
lines that became dead in a47baed.