Begin to centralize syncs-on-backend.
The doInPG method accepts 'checked' analogues of the standard functional
interfaces in java.util.function.

It also has overloads specialized for primitive return types. This isn't
strictly necessary, as the compiler is happy to add unboxing at a call
site and use the reference-returning version, but the specialized
versions allow it to avoid that.

This should allow finishing up the conversions of synchronized blocks
including those that throw checked exceptions.

In Java 8, javac has sprouted the ability to make JNI header files,
leaving less need for javah ... but not exactly *no* need, because javah
can extract the static final fields we need from the JDK class
java.sql.Types, something it's not obvious can be done with javac -h.
Happily, as of 2019, there have been no changes to the java.sql.Types
constants since Java 8; as we are now making Java 8 the minimum, just
unconditionally use the fallback java_sql_Types.h file that was
introduced in af0577a.

That saves the fuss of keeping the nar-maven-plugin's javah goal working
now that some actual Java 8 features, like lambdas, are in the class
files. The plugin tries to use BCEL to find out whether the classes
to be h'd contain anything that needs h'ing; in 3.2.3 it only considers
native methods, not @Native fields, and uses a version of BCEL that chokes
on such Java 8 features. Updating the plugin to 3.6.0 would fix the BCEL
issue, but for the simple purposes here, it's as easy to stop using javah.

It seems that the java.sql.Types constants are only used in Oid.c ...
find another way to do what that code's doing, and maybe the build can be
simplified further.

The version of maven-compiler-plugin should be updated one day; 2.5.1
only supports the (since deprecated) compilerArguments element. When it
is updated, the pom should be adjusted to use compilerArgs instead.