Verify nobody's using get() in a conflicting way
The current clear() done in scheduleJavaReleased() has a benefit:
it ensures the object won't hit the reference queue until it is
explicitly put there by our enqueue(). It can simplify the queue
drainer's job if it can know the only objects it will see are
non-RELEASED ones put there by gc, or RELEASED ones arriving only
after their last pin is gone.

But the clear(), at present, happens too late to accomplish that.
It ought to be done right away in releaseFromJava(), so it isn't
subject to queuing by gc while waiting for any pins to be gone.

That would give releaseFromJava() an immediate effect visible to
code currently holding a pin, if that code used get() expecting to
find a non-null reference there. So (brute force solution), declare
get() off limits to clients, and confirm current code doesn't break.