Sop for javac's --release 6 check
As reported in issue #235, the --release option in newer javac
versions will check against a ct.sym file included in the jdk
that is meant to record what API was available in which prior
releases, and that file (at least as shipped in Java 11) is
missing the newFactory() methods on XMLInputFactory and
XMLOutputFactory for release 6, so it reports them not found.

Happily, both no-arg newFactory() methods are equivalent to
the no-arg newInstance() methods. Deprecation tags were added
to the newInstance(String,ClassLoader) methods, but not to
the no-arg ones.