Fix a Javadoc build issue with Java 11+
As reported in #239, javadoc with an 11 or newer JDK can
fail with a message about code in modules being linked to
Java API code in the unnamed package. The Java bug report
indicates a fix was rolled out (adding code to detect when
the 'module' being detected is really an automatic module),
but it seems a <source> element in the configuration of the
maven-javadoc-plugin is still needed. That will just specify
a hardcoded 8 for now (even if the --release given to javac
will be 7 or 6); javadoc doesn't seem to mind if it is not
an exact match. (Even an actual javadoc 7 seems not to object
to source 8.)

In passing: with that hurdle cleared, javadoc 12 proceeds to
complain about non-HTML5 attributes in Deployer's javadoc, so
those are removed here (even though Deployer itself is on the
way out).