Allow Javadoc to build with Java 13
For Java 13, the version of the maven-javadoc-plugin must be
advanced to avoid pulling in a version of commons-lang whose
getJavaVersionAsFloat() method is broken by a change to the
version property values reported in 13.

There seems to be no choice of maven-javadoc-plugin version
that will work with 13 and also back to 6, so use profiles
based on the build-time JDK version to select the plugin version
to use.

A benefit of advancing the maven-javadoc-plugin version is that
the <source> element added in a28db9f can be removed again, as the
newer plugin does not need it.

Javadoc 13 is more strict in checking HTML5 structure, and disallows
<h1> in a doc comment (a level-1 header is generated by Javadoc itself),
requiring h elements to be used in proper sequence with h2 outermost.

There is nothing like a tedious bout of header renumbering to find
a few other assorted typos in doc comments.