Some better provision for release with pins
Work a bit harder to avoid having GC glom on to a JAVA_RELEASED
instance and put it on the reference queue earlier than the
releaseFromJava() logic would. Done by moving the clear() earlier,
so GC is less tempted.

There can still be cases where a gc / releaseFromJava race is
detectable but only after the fact, so also rework the queue
runner to handle the unlikely case if seen. The same approach
can cover instances that have not been JAVA_RELEASED, but simply
allowed to go unreachable with outstanding pins. For now, the
logic will only poll them and wait for the pin count to hit zero,
which is reasonable if one assumes there is still a thread somewhere
that is doing something under each pin and will shortly release it
(such a thread could have a reference to the DualState instance but
not to its referent). There's not yet any provision for questioning
that assumption, but if anything ends up taking more than one ride
on the deferredReleased queue, it will be reflected in a JMX-visible
repeatedlyDeferred counter.