An XML rendering of a pg_node_tree
A <list> whose elements are all the same type will have the
attribute "all" with a value of "int", "oid", or "bit", and its
children will all be <v> elements. Otherwise, a <list> will have
no "all" attribute, and its children will be elements whose names
are a node name (e.g. FUNCEXPR), or T_Integer, T_Float, T_String,
or T_BitString.

The (b list syntax, for a bit set, is used very commonly
(more so than i for integer or o for oid), but seems never to
have been added in PG's nodeRead() ... because the whole syntax
has drifted away from being a self-describing Lisp-like thing
and toward parsing knowledge being squirreled away inside every
node-reading function in readfuncs ... and everything there that
uses a (b list handles it directly, so no one would notice that
nodeRead() has, to this day, never been taught about it.

For this "version" of nodeRead(), (b will be recognized the same
way as (i and (o.

For a <CONST>, the <member> with name "constvalue" will also have
a "length" attribute, and the member's content will be the raw content
of the Datum as xs:hexBinary.