Introduce ByteBufferXMLReader
This is an abstract base for classes that will see a readable
ByteBuffer (windowing some PostgreSQL data type) and present
the contents in a synthesized XML form as a SyntheticXMLReader.

The chief method a subclass needs to implement is
next(ByteBuffer), which will be passed a readable buffer,
suitably pinned. This class handles the rest,
including reporting a StartDocument and an EndDocument event before
and after the synthesized ones, respectively.

The next() method returns a SyntheticXMLReader.EventCarrier, which is
a simple closure with a toSAX() method to be implemented, and several
provided methods that toSAX() can use to obtain SAX handlers and
disgorge one or more parse events onto them. In other words, next() is a
bit of an XML "pull" API (though it has wide discretion over how much to
pull at one time and hold in its closure).  This allows some control
over how much to pull at one time while holding the pin on the
PostgreSQL native state, as compared to simply grabbing the pin for all
of the SAX parse() method in one go.

At first glance, the application would seem ideal for StAX. However:

StAX is a nice API, but it turns out to be hopeless to use it as
source or result with TRAX (Transformer implementations in the JRE).
That's not really the fault of TRAX either (which originated in
Apache Xerces): its support for StAX was added later (the files
StAXEvent2SAX.java, StAXStream2SAX.java, SAX2StAXStreamWriter.java,
in the trax/ source directory, have different copyright banners and
authors, within Oracle rather than the Xerces crew) and they are rife,
rife with bugs. They are especially rife in the area of handling
'content' fragments, which the code shows some signs of having
contemplated, but no signs of ever having tested.

It seems likely that such a nest of bugs shipped in the JRE for so long
will be unlikely to get fixed and risk a behavior change that disrupts
whatever awful workarounds may have been developed in unknown client code.

The only safe verdict can be: feel free to use StAX from hand-laid code
that will use the API directly if its simplicity appeals; never, ever,
ever request a StAXSource or StAXResult with any intention of using it
with a TRAX transformation.

Therefore, SyntheticXMLReader implements the SAX API instead (despite
all of the fiddliness that involves), with EventCarrier serving as a
kind of SAX-flavored "pull" event holder. It just appears that SAX is
the older Java API and is better supported and tested with TRAX and
the rest of the Java XML APIs.