Systematize Identifier's deparse() and toString()
deparse() now takes a Charset argument, for the purpose of
verifying that the identifier can be successfully written in
the specified charset (or, for the future, the purpose of
Unicode-escaping it if necessary to make sure it can be
written; for now, it will just complain).

toString() is now exactly deparse(UTF_8).