Make Lexicals.Identifier serializable
One may want to use Identifier in a custom Principal subclass, and
any Principal has to implement Serializable (if it is to be used
in a Subject, anyway). Perhaps only Simple and its subclasses
needed to be made serializable, but the rest of the Identifier
hierarchy wasn't much harder.

Add Pseudo, intended as a principled approach to the special
identifiers that represent, say, PUBLIC as a privilege grantee.