This runs the DDRProcessor, but
... it produces a NullPointerException from deep within
com.sun.tools.javac.model.JavacTypes.getDeclaredType
when called from DDRProcessor.java:252, which is the first
attempt to get a Type for a class in java.sql (a non-java.base
module). But why? There's a requires java.sql in the module
declaration; nothing is complaining about access to the package...