The cause: javac 10 and 11 are broken
In 10 and 11, the presence of a --release 9 (for either), or
--release 10 (for 11), will cause getDeclaredType to think it
is seeing java.sql types in *all of the modules that transitively
require java.sql* ... and having seen more than one, rules it
ambiguous and refuses to see it at all. In other words, this is
something that breaks when 10 or 11 is given a --release from the
past. 10 works fine building for 10, 11 works fine building for 11.

9 builds for itself just fine, and whatever the issue was in 10 and 11
got fixed in 12, so 12 and 13 can build for --release 9 with no trouble.

This is hacky.