Dedicate an exported package to mbeans
Only the interface has to go there. It seems it must get an
unqualified export, as whatever internal machinery does the
querying shows up as some unnamed@1322ba3d-ish mystery module.

And with that, everything that worked pre-java9ify seems again
working. I was honestly expecting that to be harder.