Begin moving more of method lookup to Java.
The goal is to get the proper method looked up in Java in the form
of a MethodHandle, rather than poking details down to C and having
a jmethodID looked up there. This is only a step along.

In passing, add a specialization() method similar to the one in the
SQL generator (only longer and messier, as the API available for use
at runtime is way more fiddly than what an annotation processor can use
at compile time). Nothing uses it yet, but uses are anticipated.