OpenJ9 calls Reference.clear() from enqueue()?!
Therefore on OpenJ9 one doesn't get very far if DualState
overrides clear() to assert that no one calls it.

So, weaken the assertion to allow the call if the instance
really has been JAVA_RELEASED. This should happen nowhere
but on OpenJ9.

The behavior was added for Java 9 (but in the hg.openjdk.java.net
sources, the reference is cleared "by magic" similarly to when the
GC does it, while the variant used for OpenJ9 calls the 'clear'
virtual method).

This will have to be backpatched to 1.5, which is advertised
as usable on newer JVMs and on OpenJ9.

See https://github.com/AdoptOpenJDK/openjdk-support/issues/42

Backpatch of a6a5708.