Dedicate a non-exported package to the Processor
The SQL-generating DDR processor needs to be in the pljava-api
module so it can be provided to the java compiler, which uses
it. But its implementation really isn't part of PL/Java's API.

The Lexicals class has apparently been found useful by some
developers, so it gets to stay in the sqlgen package and be
exported. The rest of the annotation processor gets a new,
non-exported package.