Complete Checked's set of of() methods
They can be handy for writing and using lambda expressions
in places where the grammar doesn't allow them naked, or where
the compiler doesn't infer the intended type.

In the same breath, add cleaner versions of AccessController's
doPrivileged... methods. It's exhausting to use that API.

These wrapper methods must never ever ever be exported from this
module, because the real methods on AccessController will see
these wrappers as the caller, and apply this module's own privs.
But as long as they are only used within this module anyway, that
isn't a problem.

# Conflicts:
#	pljava/src/main/java/org/postgresql/pljava/internal/Checked.java