Embrace the intra-module lookup limitations
After all, it was never intended to let anybody CREATE FUNCTION
and name any method inside org.postgresql.pljava.internal!

By providing an intra-module-capable Lookup with only public
access enabled, and only when the target class is Commands,
the actually-wanted result of allowing just the public static
methods in Commands to be CREATE FUNCTION targets, follows
naturally.

There is still an addClassImages method in Commands that's
declared public just because a method in Backend calls it,
which is an implementation artifact that seems fixable.