Adapt the method handle to invoke from JNI
That's a trick headline. You can't invoke a method handle from JNI.
adaptHandle() here piles a bunch of MethodHandle combinators on top
of the underlying handle, to produce something that can be passed an
object array and a ByteBuffer. The idea is for the C code to segregate
the actual parameters, putting all those of reference type into an
object array, and the primitives into an array of jvalue mapped with
a direct byte buffer, and pass the method handle, array, and buffer
to a Java method that does the invocation.

This does seem somewhat unpleasingly roundabout.