Clear the loader cache if deployment fails
Cached references to defunct classes from a failed jar install
are not conducive to a later attempt succeeding.

This is not yet a complete solution: it is now possible
for an installJar to fail, be retried, and not bonk into
a method-not-found exception (because of old defunct classes
in the signature with the right names but wrong loaders),
but still bonk into a "cannot cast Foo to Foo" class cast
exception for the same sort of reason, if a mapped UDT is
involved, which Type is caching, I suspect.

But neither problem is new with this refactoring; they can
be triggered in the former code the same way. And the workaround
is pretty simple, just get a new backend with \c. So tracking down
the rest of this doesn't have to be a priority.