pljava.classpath => pljava.module_path
This works differently enough now that it should be reflected
in a name change for the setting. It should now have two jars
(the pljava internals jar and the pljava-api jar), separated
by the correct path separator for the platform. As before, the
default value should be right, if the jars are unpacked in the
expected places.

There may yet be uses in the field where something else was added
to pljava.classpath, and where the something-else possibly won't
work if placed on the module path. For such a case, there might
still be an argument for re-adding a GUC setting to control an
old-fashioned class path ... probably spelled pljava.class_path
to be closer to the modern Java usage and avoid confusion with the
old setting. That has not been done here either.

Even if that is not done, some thought needs to be given to what
the JVM's class-path should be set to when no class-path is desired.
It should be given a default value that prevents any classes being
found there, which is surprisingly hard. If left empty, it simply
means the current directory, and will be distractable by anything
named *.class in the PostgreSQL backend's CWD. If set to the name
of a nonexistent jar, it complains that the jar doesn't exist, and
refuses to start. It could be set to a directory guaranteed to
contain no .class files, but how to reliably arrange that?