Make empty classpath mean no classpath
Not as difficult as I had anticipated. The getSystemClassLoader apidoc
https://docs.oracle.com/javase/9/docs/api/java/lang/ClassLoader.html#getSystemClassLoader--
has this to say:

"... the system property "java.class.path" during early initialization
of the VM. If the system property is not defined, or its value is
an empty string, then there is no class path when the initial module
is a module on the application module path, i.e. a named module.
If the initial module is not on the application module path then the
class path defaults to the current working directory."

So, whether an empty java.class.path means "current directory" or
"there is no class path" comes down to whether "the initial module"
is a named module. What that means, more precisely, is here:
http://hg.openjdk.java.net/jdk9/jdk9/jdk/file/65464a307408/src/java.base/share/classes/jdk/internal/loader/ClassLoaders.java#l75
which looks for a non-null value of jdk.module.main.

Sure enough, adding a -Djdk.module.main=org.postgresql.pljava to the
JVM launch options does the trick, ensuring that the default, empty,
value of java.class.path means "there is no class path" rather than
"trust random .class files you find in the datadir".

Meanwhile, if a site for some reason wants to load some classes from
the datadir, simply adding -Djava.class.path=. will accomplish that,
just as one would naïvely expect.

As the need for a site to specify a class (as opposed to module) path
should be rare, and it can be done with -Djava.class.path= added in
pljava.vmoptions, it doesn't seem pressing to add any dedicated new
GUC for it.

In passing, just a few other doc tweaks of minor notes obsoleted
by Java 9.