Update the hello, world example
The hello pom should specify version 3.8.1 (or 3.8.0 or later,
anyway) of the maven-compiler-plugin, to avoid having it default
to -source/-target versions that newer javac versions don't
support. Even 3.8.1 only moves the default release up to 6, which
javac 12 has dropped support for. So just add an explicit release 9
to head off "why didn't that work" questions; hello, world isn't
supposed to be a problem-solving test.

For the same reason, put back the pre-JPMS service loader config
file for the DDR processor that was removed in dd91fca, so that
also Just Works; maven doesn't automagically add the API jar
to --processor-module-path.