Javadoc generation also java9ified
The maven-javadoc-plugin (at the current release, 3.1.1) still lives
in a mostly pre-JPMS world, and boy does it show. Javadoc 9 introduces
handy new options to control what should be documented for a module,
based on what is declared for export in the module descriptor, all
of which can be figured out by javadoc itself. The maven-javadoc-plugin
forces all of that not to work right, by passing an explicit list of
packages, computed the wrong way, which overrides whatever convenient
selection option was passed to javadoc.

Interestingly, though the plugin has not been updated to handle the
new options specific to javadoc 9 (they just have to be passed as
<additionalOption>s), it does have a getPackageNamesRespectingJavaModules
method, which tries to duplicate what javadoc itself would do if passed
--show-packages exported—which would be ok, more or less, as long as
'--show-packages exported' was what you wanted, and if it were done right.
Alas, it doesn't seem to understand you might export a.b and not a.b.c,
so it can end up forcing the documentation of packages that were not
intended as API.

There seems to be one (maybe only one?) roundabout way to get the plugin
not to supply the list of package names, and thereby let javadoc just do
the right stuff it can already do. A fix for a reported bug MJAVADOC-365
tweaked the plugin's executeReport method to suppress the package list
if sourceFileExcludes or sourceFileIncludes are active. So it becomes
necessary to use one of those options.

Even that must be done with care. The list of includes or excludes can't
be empty, or the trick is ignored, but also can't include too much, lest
it once again override javadoc's efforts to do the right thing.

When the includes/excludes are not used, the plugin generates an argfile
by default that names only one source file, the module-info.java. So an
obvious idea would be to add sourceFileIncludes naming only that file.

That doesn't incur any reported errors; it just prevents javadoc from
running at all. Maven simply completes all the other 'site' activities,
says nothing at all about javadoc, and calls it success.

So the other almost-as-obvious workaround idea is to add
sourceFileExcludes, and specify some file that wasn't going to be
included anyway. But that causes a full enumeration of the other source
files to be generated and passed to javadoc, overriding its efforts to
do the right thing.

So it becomes necessary to use sourceFileExcludes with a wildcard pattern
to exclude everything (except the module-info.java file), producing the
same list of files (just one file) that was produced by default, while
also suppressing the list of packages. The pattern used here just excludes
all .java files under org/ (which is where all of them are, except for
module-info.java at top level).

That does the needed trick, but ends up leaving javadoc not finding any
classes to document at all; that is easily fixed with a --module option.

In passing, clean up three errors in javadoc comments that were flagged
by javadoc, once the excitement of just getting it to work died down,
and fix a comment that still referred to Java pre-9.