Zero the static parameter slots after use
Zero/null the static slots after the parameter values have been
loaded onto the stack for the target method. For both primitives
and references, this is worth doing to avoid having potentially
sensitive values indefinitely sitting in the static area. For
references, it's also important to avoid indefinitely holding
them live.

It seems to add maybe 100 ns to null a reference, and negligible
cost to zero a primitive.