Allow rare reentrant invocation activity
The preparation to invoke a general PL/Java function involves
converting parameters and storing them in a static area from which
the function's Java method handle will fetch them. The fetching is
complete and they are all on the Java stack before the target method
is invoked, so reentrant PL/Java invocations from that method do not
pose any problem.

Reentrant invocation while enumerating and converting those parameters
can pose a problem. It could occur if one of the parameters is a PL/Java
UDT. The UDT read/write/parse/toString methods themselves are handled
specially, without touching the static parameter area, so they do not
directly exercise this case. But no rule limits what a PL/Java UDT
implementing method can do, so should one involve a query that calls
another PL/Java function, the problem case arises.

It can be detected by seeing that the countCheck field in the parameter
area is already nonzero, saving the slot contents if so, and restoring
them when the current invocation (normally or exceptionally) completes.
To make the countCheck test reliable, every general PL/Java function's
MethodHandle tree is now constructed to leave countCheck zero after
all the parameters have been fetched.

Adds a bit of copying overhead in the expected-to-be-rare case, but
keeps it out of the common case.

Add a few lines to an existing UDT example to make sure this gets
exercised.