A start on a function validator
PL/Java has been without one long enough. This start does not include
much of an effort to improve or beautify the existing error messages for
unsuccessful function definitions, but only to make the same old ugly
ones happen at a more useful time, that of CREATE FUNCTION instead of
only later with attempts to use the function.

The validator is shoehorned in by passing a forValidator boolean to
Function_getFunction, which is now capable of returning NULL, rather
than always a usable Function instance. Naturally, it should only
return NULL when forValidator is true. The validator doesn't use
the result.

For a non-UDT function, a complete, usable Function should result
if check_function_bodies is true, but not if it's false. If it is
produced, it is saved in the hash table; no point re-doing the work
later when the function is used.

For a UDT function, incomplete initialization will be the rule,
not the exception, as UDT functions are created while the associated
PostgreSQL type is only a shell, and that's too early to save any
vitals about it.